refactor(carbon): honour the layering rule for the access guard
CLAUDE.md 第 1 條:唯 repository 可存取 Prisma,唯 service 可使用 repository。
本分支新增的 carbon_access 同時違反兩者 —— 它位於 lib/,卻既直接呼叫
prisma.chatroom.findUnique 又引用 accountBookRepo。

1. lib/carbon_access.ts → services/carbon_access.guard.ts
   命名比照既有的 services/account_book_access.guard.ts:它做的是授權裁決,
   本質是 service 而非工具函式。搬到 services/ 後引用 repository 才合法。

2. prisma 直呼改為 chatroomRepo.findAccountBookIdByChannel()
   該方法早已存在且語意完全相同(where channel → select accountBookId →
   ?? null),屬直接替換,無行為變更。

3. sessions/route.ts 不再直接引用 accountBookRepo
   原本在 route 內做 getMemberRoleByAddress 授權判斷,改用同模組既有的
   canViewAccountBook()。該路由本來就已 import 這個模組,等於在旁邊重寫了
   它的內容。
   注意兩者參數順序相反:canViewAccountBook(userAddress, accountBookId)
   而 repo 是 getMemberRoleByAddress(accountBookId, userAddress)。

4. carbon_access.test.ts 的 mock 一併改為 @/repositories/chatroom.repo
   原本 mock @/lib/prisma 並操作 prisma.chatroom.findUnique;實作改走
   repository 後該 mock 不再攔截到目標,不改會讓測試失去意義。
   回傳型別由 { accountBookId } 物件改為 string | null。
